requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
psycopg2-binary
python-jose
python-multipart
pyjwt
alembic
fastapi-login
